admin access added

+++ src/app/api/users/[id]/route.jsx
@@ -1,6 +1,7 @@
 import { NextResponse } from "next/server";
 import {doc,getDoc,deleteDoc} from 'firebase/firestore'
 import { db } from "@/app/firebase";
+import { auth } from "@/auth";
 
 //search for a specific user
 export async function GET(_,res){
@@ -24,31 +25,38 @@ export async function GET(_,res){
 }
 
 // login
-export async function POST(req, res) {
-    const { name, email, password } = await req.json();
-    const { id } = res.params;
-    const { name: newName, email: newEmail, password: newPassword } = usersData.find((user) => id === user.id);
+// export async function POST(req, res) {
+//     const { name, email, password } = await req.json();
+//     const { id } = res.params;
+//     const { name: newName, email: newEmail, password: newPassword } = usersData.find((user) => id === user.id);
 
-    if (newName === name && newEmail === email && newPassword === password) {
-        return NextResponse.json({ result: "successfully logged in." });
-    }
-    else if (!name || !email || !password) {
-        return NextResponse.json({ result: "please fill all the input fields" });
-    }
-    return NextResponse.json({ result: "Invalid user credentials" });
-}
+//     if (newName === name && newEmail === email && newPassword === password) {
+//         return NextResponse.json({ result: "successfully logged in." });
+//     }
+//     else if (!name || !email || !password) {
+//         return NextResponse.json({ result: "please fill all the input fields" });
+//     }
+//     return NextResponse.json({ result: "Invalid user credentials" });
+// }
 
 //deleting a specific user
 
 export async function DELETE(req,res){
     const {id}=res.params;
+    const session=await auth();
+    
+    if(session?.user?.email==="admin@gmail.com"){
     // console.log(id);
-    try {
-        const userDoc = doc(db, 'users', id);
-        await deleteDoc(userDoc);
-        return NextResponse.json({ result:"user data deleted successfully"},{status:200});
-    } catch (error) {
-        console.log(error.message);
-        return NextResponse.json({result:"user not found"},{status:404});
+        try {
+            const userDoc = doc(db, 'users', id);
+            await deleteDoc(userDoc);
+            return NextResponse.json({ result:"user data deleted successfully"},{status:200});
+        } catch (error) {
+            console.log(error.message);
+            return NextResponse.json({result:"user not found"},{status:404});
+        }
+    }
+    else{
+        return NextResponse.json({result:"access unavailable"},{status:500});
     }
 }
